Update ControllerAdvice handlers with ErrorMessageConfig injections
