+++ src/main/java/com/example/echo_api/controller/account/AccountControllerAdvice.java
@@ -1,31 +1,12 @@
 package com.example.echo_api.controller.account;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 
 import com.example.echo_api.exception.AbstractControllerAdvice;
-import com.example.echo_api.exception.custom.password.PasswordException;
-import com.example.echo_api.exception.custom.username.UsernameException;
-import com.example.echo_api.persistence.dto.response.error.ErrorResponse;
-
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @ControllerAdvice(assignableTypes = AccountController.class)
 public class AccountControllerAdvice extends AbstractControllerAdvice {
-
-    @ExceptionHandler({ UsernameException.class, PasswordException.class })
-    ResponseEntity<ErrorResponse> handleUsernamePasswordException(HttpServletRequest request, Exception ex) {
-        log.debug("Handling exception: {}", ex.getMessage());
-
-        return createExceptionHandler(
-            request,
-            HttpStatus.BAD_REQUEST,
-            ex.getMessage(),
-            null);
-    }
 
 }
