+++ src/main/java/com/example/echo_api/exception/GlobalControllerAdvice.java
@@ -13,6 +13,9 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
 
+import com.example.echo_api.config.ErrorMessageConfig;
+import com.example.echo_api.exception.custom.password.PasswordException;
+import com.example.echo_api.exception.custom.username.UsernameException;
 import com.example.echo_api.persistence.dto.response.error.ErrorResponse;
 
 import jakarta.servlet.http.HttpServletRequest;
@@ -27,6 +30,12 @@ import lombok.extern.slf4j.Slf4j;
  * The methods in the class handle global occurrences for common HTTP status
  * codes (401, 403, 404, 500) across all controllers.
  * 
+ * <p>
+ * Methods also include Jakarta validation exception handlers.
+ * 
+ * <p>
+ * Methods also include {@link UsernameException} and {@link PasswordException}
+ * abstract class handlers.
  */
 @Slf4j
 @ControllerAdvice
@@ -56,7 +65,7 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
         return createExceptionHandler(
             request,
             HttpStatus.BAD_REQUEST,
-            "invalid request",
+            ErrorMessageConfig.INVALID_REQUEST,
             details);
     }
 
@@ -83,8 +92,20 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
         return createExceptionHandler(
             request,
             HttpStatus.BAD_REQUEST,
-            "invalid request",
+            ErrorMessageConfig.INVALID_REQUEST,
             details);
+    }
+
+    /** Username/Password Exception */
+    @ExceptionHandler({ UsernameException.class, PasswordException.class })
+    ResponseEntity<ErrorResponse> handleUsernamePasswordException(HttpServletRequest request, Exception ex) {
+        log.debug("Handling exception: {}", ex.getMessage());
+
+        return createExceptionHandler(
+            request,
+            HttpStatus.BAD_REQUEST,
+            ex.getMessage(),
+            null);
     }
 
     /** 401 */
@@ -95,7 +116,7 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
         return createExceptionHandler(
             request,
             HttpStatus.UNAUTHORIZED,
-            "unauthorised request",
+            ErrorMessageConfig.UNAUTHORISED,
             null);
     }
 
@@ -107,7 +128,7 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
         return createExceptionHandler(
             request,
             HttpStatus.FORBIDDEN,
-            "access denied",
+            ErrorMessageConfig.FORBIDDEN,
             null);
     }
 
@@ -119,7 +140,7 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
         return createExceptionHandler(
             request,
             HttpStatus.NOT_FOUND,
-            "resource not found",
+            ErrorMessageConfig.NOT_FOUND,
             null);
     }
 
@@ -130,7 +151,7 @@ public class GlobalControllerAdvice extends AbstractControllerAdvice {
 
         return createExceptionHandler(request,
             HttpStatus.INTERNAL_SERVER_ERROR,
-            "an unexpected error occurred",
+            ErrorMessageConfig.INTERNAL_SERVER_ERROR,
             ex.getMessage());
     }
